restructure profile routes for role-based access
