+++ src/app/api/notifications/route.ts
@@ -3,70 +3,60 @@ import { getServerSession } from 'next-auth';
 import connectDB from '@/lib/mongodb';
 import Notification from '@/models/Notification';
 import { authOptions } from '@/lib/auth';
+import { asyncHandler, AuthenticationError, createSuccessResponse } from '@/lib/errorHandling';
 
 // GET notifications for current user
-export async function GET(req: NextRequest) {
-  try {
-    const session = await getServerSession(authOptions);
-    if (!session) {
-      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-    }
-
-    await connectDB();
-
-    const notifications = await Notification.find({ userId: session.user?.id })
-      .populate('billId', 'roomId month year totalAmount')
-      .sort({ sentAt: -1 })
-      .limit(50);
-
-    const unreadCount = await Notification.countDocuments({
-      userId: session.user?.id,
-      read: false
-    });
-
-    return NextResponse.json({
-      success: true,
-      data: notifications,
-      unreadCount
-    });
-  } catch (error) {
-    console.error('Error fetching notifications:', error);
-    return NextResponse.json({ error: 'Failed to fetch notifications' }, { status: 500 });
+export const GET = asyncHandler(async (req: NextRequest) => {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    throw new AuthenticationError('Unauthorized');
   }
-}
+
+  await connectDB();
+
+  const notifications = await Notification.find({ userId: session.user?.id })
+    .populate('billId', 'roomId month year totalAmount')
+    .sort({ sentAt: -1 })
+    .limit(50);
+
+  const unreadCount = await Notification.countDocuments({
+    userId: session.user?.id,
+    read: false
+  });
+
+  return createSuccessResponse({
+    notifications,
+    unreadCount
+  });
+});
 
 // POST mark notification as read
-export async function POST(req: NextRequest) {
-  try {
-    const session = await getServerSession(authOptions);
-    if (!session) {
-      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-    }
-
-    const { notificationId } = await req.json();
-
-    if (!notificationId) {
-      return NextResponse.json(
-        { error: 'notificationId is required' },
-        { status: 400 }
-      );
-    }
-
-    await connectDB();
-
-    const notification = await Notification.findOneAndUpdate(
-      { _id: notificationId, userId: session.user?.id },
-      { read: true, readAt: new Date() },
-      { new: true }
-    );
-
-    if (!notification) {
-      return NextResponse.json({ error: 'Notification not found' }, { status: 404 });
-    }
-
-    return NextResponse.json({ success: true, data: notification });
-  } catch (error) {
-    console.error('Error updating notification:', error);
-    return NextResponse.json({ error: 'Failed to mark as read' }, { status: 500 });
+export const POST = asyncHandler(async (req: NextRequest) => {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    throw new AuthenticationError('Unauthorized');
   }
-}
+
+  const { notificationId } = await req.json();
+
+  if (!notificationId) {
+    return NextResponse.json(
+      { error: 'notificationId is required' },
+      { status: 400 }
+    );
+  }
+
+  await connectDB();
+
+  const notification = await Notification.findOneAndUpdate(
+    { _id: notificationId, userId: session.user?.id },
+    { read: true, readAt: new Date() },
+    { new: true }
+  );
+
+  if (!notification) {
+    return NextResponse.json({ error: 'Notification not found' }, { status: 404 });
+  }
+
+  return createSuccessResponse(notification, 'Notification marked as read');
+});
